refactor: add dynamicCss function to avoid loading editor css from preferences module
